chore: radio support wave
