+++ components/_util/wave/interface.ts
@@ -1,0 +1,15 @@
+import type { GlobalToken } from '../../theme';
+
+export const TARGET_CLS = 'ant-wave-target';
+
+export type ShowWaveEffect = (
+  element: HTMLElement,
+  info: {
+    className: string;
+    token: GlobalToken;
+    component?: string;
+    event: MouseEvent;
+  },
+) => void;
+
+export type ShowWave = (event: MouseEvent) => void;
